[TWINFACES-944] feat: update PR
