+++ src/app/core/historyes/page.tsx
@@ -1,0 +1,5 @@
+import { HistoryesScreen } from "@/screens/historyes";
+
+export default function Page() {
+  return <HistoryesScreen />;
+}
