+++ src/screens/historyes/historyes.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { HistoryesTable } from "@/widgets/tables";
+
+export function HistoryesScreen() {
+  return <HistoryesTable />;
+}
