+++ src/screens/twin/views/index.ts
@@ -2,6 +2,5 @@ export * from "./twin-attachments";
 export * from "./twin-comments";
 export * from "./twin-fields";
 export * from "./twin-general";
-export * from "./twin-history";
 export * from "./twin-links";
 export * from "./twin-trigger-tasks";
