+++ src/widgets/layout/sidebar-layout/sidebar/constants.ts
@@ -6,6 +6,7 @@ import {
   BriefcaseBusiness,
   Eraser,
   Factory,
+  History,
   Inbox,
   Layers,
   Link2,
@@ -102,6 +103,11 @@ export const SIDEBAR_GROUPS: Record<GroupKeys, Group> = {
         title: "Attachments",
         url: `/${PlatformArea.core}/attachments`,
         icon: Paperclip,
+      },
+      {
+        title: "History",
+        url: `/${PlatformArea.core}/historyes`,
+        icon: History,
       },
     ],
   },
