+++ src/widgets/tables/historyes/index.ts
@@ -1,0 +1,1 @@
+export * from "./historyes";
