+++ src/widgets/tables/index.ts
@@ -5,6 +5,7 @@ export * from "./factory-conditions";
 export * from "./factory-multipliers";
 export * from "./factory-pipelines";
 export * from "./factory-triggers";
+export * from "./historyes";
 export * from "./permission-groups";
 export * from "./permissions";
 export * from "./pipeline-steps";
